Rename main chartData module as index.js
This way it is more clear what is the importable file for the outside
modules. Other modules should not import `getViewerData`, `parseUtils`
or `tree` files, as they should be considered internal to the
getChartData function.
